unit tests for non-torrent reports

+++ app/Stats/Report.php
@@ -3,7 +3,6 @@
 namespace Gazelle\Stats;
 
 class Report extends \Gazelle\Base {
-
     public function day(int $day): array {
         self::$db->prepared_query("
             SELECT r.ResolverID AS user_id,
@@ -45,7 +44,7 @@ class Report extends \Gazelle\Base {
     public function torrentOpenTotal(): int {
         $count = self::$cache->get_value('num_torrent_reportsv2');
         if ($count === false) {
-            $count = self::$db->scalar("
+            $count = (int)self::$db->scalar("
                 SELECT count(*) FROM reportsv2 WHERE Status = 'New'
             ");
             self::$cache->cache_value('num_torrent_reportsv2', $count, 3600);
@@ -56,7 +55,7 @@ class Report extends \Gazelle\Base {
     public function otherOpenTotal(): int {
         $count = self::$cache->get_value('num_other_reports');
         if ($count === false) {
-            $count = self::$db->scalar("
+            $count = (int)self::$db->scalar("
                 SELECT count(*) FROM reports WHERE Status = 'New'
             ");
             self::$cache->cache_value('num_other_reports', $count, 3600);
@@ -67,7 +66,7 @@ class Report extends \Gazelle\Base {
     public function forumOpenTotal(): int {
         $count = self::$cache->get_value('num_forum_reports');
         if ($count === false) {
-            $count = self::$db->scalar("
+            $count = (int)self::$db->scalar("
                 SELECT count(*)
                 FROM reports
                 WHERE Status = 'New'
